TestPilot/requirements.txt
flask
google-generativeai
zhipuai
openai
requests
